30 Modifiquei a tela de Editar Produto para receber fornecedor.

+++ src/model/Produto.java
@@ -77,7 +77,7 @@ public class Produto implements Serializable {
         this.precoVenda = precoVenda;
     }
 
-    public double getQuantidadeEstoque() {
+    public Integer getQuantidadeEstoque() {
         return quantidadeEstoque;
     }
